Add more processor boilerplate to MakeGoalFactoryProcessor
This pulls out shared code from MakeLogicValueProcessor into new modules, and
also includes the addition of the MakeGoalFactoryMetadata object, which is
basically complete already, so this is more than just boilerplate.

This also replaces some thrown exceptions with proper compiler messages. I'd
rather not develop the bad habit of throwing random exceptions as messages to
the user. I've gone ahead and printed the stack trace of IOExceptions because
those should rarely happen, and the standard Java API apparently doesn't want me
to send it anywhere besides stderr since there is no standard way to get the
stack trace in a String.