Add a note about the rawtypes warning in MakeLogicValue doc
I didn't notice this gotcha until I ran Eclipse myself and saw the warning. It
seems that rawtypes is not a very useful warning - -Xlint:unchecked should show
everything that is not safe. Whether you use "<?, ?>" or not should not change
the fact that all accessors will return Object references.