Make the two processors play nicely together
This change forces the user to make the @MakeLogicValue-annotated type generic,
which means that it can be used in pattern-matching clause parameters. Using
this works around the limitation that javac doesn't give annotation processors
any information about generic ErrorType TypeMirrors (the qualified name is
"<any>"). This means that we can now compile MakeGoalFactory and MakeLogicValue
annotations in the same target.

I considered also requiring @MakeLogicValue classes to completely hide the
implementing class, which means that there is no need to decide if one should
reference ICons<> or Cons<> at any time, as other classes would always reference
the interface/superclass. I am still considering this.

It seems that unrelated errors in certain places causes javac to throw away the
results of annotation processors, which means that you get spurious errors about
undeclared classes which should have been generated, which obscures the actual
problem which caused the other errors. This would be a good reason to "hide" the
implementing class as described above. Hopefully javac fixes this issue in
future JDK updates.