Make logic values define static factory methods
, and impose other reasonable regulations on how to use LogicValue types.
Generated LogicValue implementations are no longer public, which forces owners
of an API to hide the implementation detail. It is too confusing to make both
types public and force the user to decide which one to use. The pattern before
of using Foo (implementation) and IFoo (interface) actually made the code
cluttered and didn't feel like real Java.

The generated implementation class now has a non-customizeable name that is
generated using AutoValue's algorithm (see
https://github.com/google/auto/tree/master/value).