Add boilerplate for the processor for @MakePredicates
The first two processors followed this pattern, as well as other unrelated
processors I've worked on in the past, so I'm pretty sure it will work for this
processor as well.

This doesn't do anything cool but it makes the remaining tasks seem more
manageable. I also have a holding place to put notes and code as I think of
them, so I can drop it somewhere appropriate before I forget any of it.