Convert DiffList goals to use @MakePredicates
This looks much nicer than @MakeGoalFactory, so I'm feeling pretty good about
the migration at this point. There is a small amount of added boilerplate, but
the API is much easier to document for the developer and much easier to
understand for the user, since the user need not understand Java code
generation.