Add a simple relational Lisp interpreter
This interpreter doesn't support branching or interesting functions like
"equal". I'm not sure how I want to implement this yet. Constraints like =/= as
done in miniKanren would do the trick, but I wonder if I can do this by keeping
a list of all available atoms in the Lisp class, and using that to make things
not equal.