Add Dumper functionality for debugging
This feature was prompted by my inability to visualize the data structures and
Goal-generation process, and my difficulty in debugging the Lisp interpreter.

This feature allows generating a tree view of some nested data structure. It is
particularly important for visualizing Goal graphs and Stream state. Now you
can see (for instance) how DelayedGoals and ImmatureStreams expand into Substs
and SolveSteps over time.

I took this opportunity to also move RepeatedGoal into a separate, top-level
class, rather than an anonymous class. This makes it consistent with
DelayedGoal. Since it is not common to use RepeatedGoal in practice, the extra
typing of "new RepeatedGoal(" vs. "repeat(" is not a concern.