Allow testing for compiler errors generated by processors
Adds a test for an error in MakeGoalFactoryProcessor. It intercepts all messages
with a ClassGeneratingMessager which saves the messages to a class which can
spit out the messages at run time. These messages are read by unit tests.

This also fixes a NPE bug in MakeGoalFactoryProcessor which happened when no
valid clause methods are in an annotated class.