Show Goal before rest of stream in ImmatureStream(bind) dumps
The Goal is often simple, while the rest of the stream can be massive. Put the
simple thing first to make the dump easier to read, so the cascading is more
visually consistent.

Later, I will make the ImmatureStream cascading automatically collapse into a
single level so dumps will be even cleaner.