Remove "(FooBar)" after "ImmatureStream" in dump headings
There are three kind of ImmatureStreams and all three can be distinguished from
one another by the type of the sub-components. IS(DelayedGoal) is the only one
with a Subst in the second place. IS(bind) and IS(mplus) are distinguished from
one another by the type of the first component: a Stream and Goal, respectively.