Make Subst class an actual Map
Making Subst an actual Map — rather than simply contain a map field — makes it
easier to reason about its API.

The type of the Map key is now Object, since I want Subst to be able to contain
other kinds of data about the calculation in progress. For now, I've removed the
ability to sort the Vars predictably, since in practice this never seemed to
help me, and there weren't even any unit tests that checked for it.