CachedGoal v1
I plan to use this for implementing a DisjGoal->ConjGoal converter. To do that,
you have to turn:

(A & B) | C

into:

(A | C) & (B | C)

To make the latter work equivalently, (1) C cannot evaluate differently the two
times it is run, and (2) B should not be run if A fails. This first commit
handles (1).

For (1), if C can succeed in multiple ways, it should succeed in the same way in
(A | C) as it does in (B | C), or else extra substitutions will be generated
which combine the correct substitutions.

For (2), if B is evaluated when A fails, then (B | C) will succeed more times
than it ought to, and there will be junk substitutions in the stream where B
pointlessly adds bindings. To make this work, CachedGoal will also support
prerequisite CachedGoals, which means the delegate Goal will not be attempted
if some other CachedGoal is not already solved.

Additionally, though not a requirement for correctness, (3) we should aim for
that in (A | C) & (B | C), if C fails in (A | C), it is not re-evaluated in
(B | C), in case C is very expensive. This can be accomplished by replacing
(B | C) with (B | C'), where C' is Goals.UNIT with C as a prerequisite.