Make the Stream API a little easier to understand
I didn't like that Stream returned a SolveStep rather than a Subst for its head.
It felt like a cons list that had `car' values of alternating natures -
alternating between Subst and Stream. Also, I have another idea for refactoring
this API. It may not work out at all, but it will be easier to try with this
simplification.