Make Streams return EmptyStream in place of null in rest()
It was confusing that sometimes streams returned null, and sometimes
EmptyStream, from rest(). We use EmptyStream rather than null because it is
easier to use when mplus'ing and bind'ing streams.

Because the EmptyStream instance itself signals the end of a stream in tests,
rather than null, we now find the end of the stream faster in LogicAsserter,
which reduces the so-called workUnits for each assertion by 1.