Make Dumpable interface match the requirements of ImmatureStream
ImmatureStream had some glue code in it to allow ImmatureStreams to run longer
in dumps rather than become nested. This commit alters the API of Dumpable to
match the requirements of ImmatureStream, and move the glue code to Dumpable.

This means there are now two methods in Dumpable rather than one, and some the
code required to implement Dumpable is a little more verbose. However, it will
make the Dumper/Dumpable API more extensible, and I have a new feature idea I
want to implement some day...